+++ packages/video_player/video_player_android/android/src/main/kotlin/io/flutter/plugins/videoplayer/Messages.kt
@@ -490,13 +490,58 @@ data class PlatformVideoViewCreationParams(val playerId: Long) {
   }
 }
 
+/**
+ * Pigeon equivalent of video_player_android's WidevineDrmConfiguration.
+ *
+ * Generated class from Pigeon that represents data sent in messages.
+ */
+data class PlatformWidevineDrmConfiguration(
+    val licenseUri: String,
+    val licenseHeaders: Map<String, String>
+) {
+  companion object {
+    fun fromList(pigeonVar_list: List<Any?>): PlatformWidevineDrmConfiguration {
+      val licenseUri = pigeonVar_list[0] as String
+      val licenseHeaders = pigeonVar_list[1] as Map<String, String>
+      return PlatformWidevineDrmConfiguration(licenseUri, licenseHeaders)
+    }
+  }
+
+  fun toList(): List<Any?> {
+    return listOf(
+        licenseUri,
+        licenseHeaders,
+    )
+  }
+
+  override fun equals(other: Any?): Boolean {
+    if (other == null || other.javaClass != javaClass) {
+      return false
+    }
+    if (this === other) {
+      return true
+    }
+    val other = other as PlatformWidevineDrmConfiguration
+    return MessagesPigeonUtils.deepEquals(this.licenseUri, other.licenseUri) &&
+        MessagesPigeonUtils.deepEquals(this.licenseHeaders, other.licenseHeaders)
+  }
+
+  override fun hashCode(): Int {
+    var result = javaClass.hashCode()
+    result = 31 * result + MessagesPigeonUtils.deepHash(this.licenseUri)
+    result = 31 * result + MessagesPigeonUtils.deepHash(this.licenseHeaders)
+    return result
+  }
+}
+
 /** Generated class from Pigeon that represents data sent in messages. */
 data class CreationOptions(
     val uri: String,
     val formatHint: PlatformVideoFormat? = null,
     val httpHeaders: Map<String, String>,
     val userAgent: String? = null,
-    val backBufferDurationMs: Long? = null
+    val backBufferDurationMs: Long? = null,
+    val widevineDrm: PlatformWidevineDrmConfiguration? = null
 ) {
   companion object {
     fun fromList(pigeonVar_list: List<Any?>): CreationOptions {
@@ -505,7 +550,9 @@ data class CreationOptions(
       val httpHeaders = pigeonVar_list[2] as Map<String, String>
       val userAgent = pigeonVar_list[3] as String?
       val backBufferDurationMs = pigeonVar_list[4] as Long?
-      return CreationOptions(uri, formatHint, httpHeaders, userAgent, backBufferDurationMs)
+      val widevineDrm = pigeonVar_list[5] as PlatformWidevineDrmConfiguration?
+      return CreationOptions(
+          uri, formatHint, httpHeaders, userAgent, backBufferDurationMs, widevineDrm)
     }
   }
 
@@ -516,6 +563,7 @@ data class CreationOptions(
         httpHeaders,
         userAgent,
         backBufferDurationMs,
+        widevineDrm,
     )
   }
 
@@ -531,7 +579,8 @@ data class CreationOptions(
         MessagesPigeonUtils.deepEquals(this.formatHint, other.formatHint) &&
         MessagesPigeonUtils.deepEquals(this.httpHeaders, other.httpHeaders) &&
         MessagesPigeonUtils.deepEquals(this.userAgent, other.userAgent) &&
-        MessagesPigeonUtils.deepEquals(this.backBufferDurationMs, other.backBufferDurationMs)
+        MessagesPigeonUtils.deepEquals(this.backBufferDurationMs, other.backBufferDurationMs) &&
+        MessagesPigeonUtils.deepEquals(this.widevineDrm, other.widevineDrm)
   }
 
   override fun hashCode(): Int {
@@ -541,6 +590,7 @@ data class CreationOptions(
     result = 31 * result + MessagesPigeonUtils.deepHash(this.httpHeaders)
     result = 31 * result + MessagesPigeonUtils.deepHash(this.userAgent)
     result = 31 * result + MessagesPigeonUtils.deepHash(this.backBufferDurationMs)
+    result = 31 * result + MessagesPigeonUtils.deepHash(this.widevineDrm)
     return result
   }
 }
@@ -977,27 +1027,32 @@ private open class MessagesPigeonCodec : StandardMessageCodec() {
         }
       }
       137.toByte() -> {
-        return (readValue(buffer) as? List<Any?>)?.let { CreationOptions.fromList(it) }
+        return (readValue(buffer) as? List<Any?>)?.let {
+          PlatformWidevineDrmConfiguration.fromList(it)
+        }
       }
       138.toByte() -> {
-        return (readValue(buffer) as? List<Any?>)?.let { TexturePlayerIds.fromList(it) }
+        return (readValue(buffer) as? List<Any?>)?.let { CreationOptions.fromList(it) }
       }
       139.toByte() -> {
-        return (readValue(buffer) as? List<Any?>)?.let { PlaybackState.fromList(it) }
+        return (readValue(buffer) as? List<Any?>)?.let { TexturePlayerIds.fromList(it) }
       }
       140.toByte() -> {
-        return (readValue(buffer) as? List<Any?>)?.let { AudioTrackMessage.fromList(it) }
+        return (readValue(buffer) as? List<Any?>)?.let { PlaybackState.fromList(it) }
       }
       141.toByte() -> {
-        return (readValue(buffer) as? List<Any?>)?.let { ExoPlayerAudioTrackData.fromList(it) }
+        return (readValue(buffer) as? List<Any?>)?.let { AudioTrackMessage.fromList(it) }
       }
       142.toByte() -> {
-        return (readValue(buffer) as? List<Any?>)?.let { NativeAudioTrackData.fromList(it) }
+        return (readValue(buffer) as? List<Any?>)?.let { ExoPlayerAudioTrackData.fromList(it) }
       }
       143.toByte() -> {
-        return (readValue(buffer) as? List<Any?>)?.let { ExoPlayerVideoTrackData.fromList(it) }
+        return (readValue(buffer) as? List<Any?>)?.let { NativeAudioTrackData.fromList(it) }
       }
       144.toByte() -> {
+        return (readValue(buffer) as? List<Any?>)?.let { ExoPlayerVideoTrackData.fromList(it) }
+      }
+      145.toByte() -> {
         return (readValue(buffer) as? List<Any?>)?.let { NativeVideoTrackData.fromList(it) }
       }
       else -> super.readValueOfType(type, buffer)
@@ -1038,36 +1093,40 @@ private open class MessagesPigeonCodec : StandardMessageCodec() {
         stream.write(136)
         writeValue(stream, value.toList())
       }
-      is CreationOptions -> {
+      is PlatformWidevineDrmConfiguration -> {
         stream.write(137)
         writeValue(stream, value.toList())
       }
-      is TexturePlayerIds -> {
+      is CreationOptions -> {
         stream.write(138)
         writeValue(stream, value.toList())
       }
-      is PlaybackState -> {
+      is TexturePlayerIds -> {
         stream.write(139)
         writeValue(stream, value.toList())
       }
-      is AudioTrackMessage -> {
+      is PlaybackState -> {
         stream.write(140)
         writeValue(stream, value.toList())
       }
-      is ExoPlayerAudioTrackData -> {
+      is AudioTrackMessage -> {
         stream.write(141)
         writeValue(stream, value.toList())
       }
-      is NativeAudioTrackData -> {
+      is ExoPlayerAudioTrackData -> {
         stream.write(142)
         writeValue(stream, value.toList())
       }
-      is ExoPlayerVideoTrackData -> {
+      is NativeAudioTrackData -> {
         stream.write(143)
         writeValue(stream, value.toList())
       }
-      is NativeVideoTrackData -> {
+      is ExoPlayerVideoTrackData -> {
         stream.write(144)
+        writeValue(stream, value.toList())
+      }
+      is NativeVideoTrackData -> {
+        stream.write(145)
         writeValue(stream, value.toList())
       }
       else -> super.writeValue(stream, value)