+++ packages/video_player/video_player_android/lib/src/messages.g.dart
@@ -359,6 +359,48 @@ class PlatformVideoViewCreationParams {
   int get hashCode => _deepHash(<Object?>[runtimeType, ..._toList()]);
 }
 
+/// Pigeon equivalent of video_player_android's WidevineDrmConfiguration.
+class PlatformWidevineDrmConfiguration {
+  PlatformWidevineDrmConfiguration({required this.licenseUri, required this.licenseHeaders});
+
+  String licenseUri;
+
+  Map<String, String> licenseHeaders;
+
+  List<Object?> _toList() {
+    return <Object?>[licenseUri, licenseHeaders];
+  }
+
+  Object encode() {
+    return _toList();
+  }
+
+  static PlatformWidevineDrmConfiguration decode(Object result) {
+    result as List<Object?>;
+    return PlatformWidevineDrmConfiguration(
+      licenseUri: result[0]! as String,
+      licenseHeaders: (result[1]! as Map<Object?, Object?>).cast<String, String>(),
+    );
+  }
+
+  @override
+  // ignore: avoid_equals_and_hash_code_on_mutable_classes
+  bool operator ==(Object other) {
+    if (other is! PlatformWidevineDrmConfiguration || other.runtimeType != runtimeType) {
+      return false;
+    }
+    if (identical(this, other)) {
+      return true;
+    }
+    return _deepEquals(licenseUri, other.licenseUri) &&
+        _deepEquals(licenseHeaders, other.licenseHeaders);
+  }
+
+  @override
+  // ignore: avoid_equals_and_hash_code_on_mutable_classes
+  int get hashCode => _deepHash(<Object?>[runtimeType, ..._toList()]);
+}
+
 class CreationOptions {
   CreationOptions({
     required this.uri,
@@ -366,6 +408,7 @@ class CreationOptions {
     required this.httpHeaders,
     this.userAgent,
     this.backBufferDurationMs,
+    this.widevineDrm,
   });
 
   String uri;
@@ -378,8 +421,10 @@ class CreationOptions {
 
   int? backBufferDurationMs;
 
+  PlatformWidevineDrmConfiguration? widevineDrm;
+
   List<Object?> _toList() {
-    return <Object?>[uri, formatHint, httpHeaders, userAgent, backBufferDurationMs];
+    return <Object?>[uri, formatHint, httpHeaders, userAgent, backBufferDurationMs, widevineDrm];
   }
 
   Object encode() {
@@ -394,6 +439,7 @@ class CreationOptions {
       httpHeaders: (result[2]! as Map<Object?, Object?>).cast<String, String>(),
       userAgent: result[3] as String?,
       backBufferDurationMs: result[4] as int?,
+      widevineDrm: result[5] as PlatformWidevineDrmConfiguration?,
     );
   }
 
@@ -410,7 +456,8 @@ class CreationOptions {
         _deepEquals(formatHint, other.formatHint) &&
         _deepEquals(httpHeaders, other.httpHeaders) &&
         _deepEquals(userAgent, other.userAgent) &&
-        _deepEquals(backBufferDurationMs, other.backBufferDurationMs);
+        _deepEquals(backBufferDurationMs, other.backBufferDurationMs) &&
+        _deepEquals(widevineDrm, other.widevineDrm);
   }
 
   @override
@@ -859,29 +906,32 @@ class _PigeonCodec extends StandardMessageCodec {
     } else if (value is PlatformVideoViewCreationParams) {
       buffer.putUint8(136);
       writeValue(buffer, value.encode());
-    } else if (value is CreationOptions) {
+    } else if (value is PlatformWidevineDrmConfiguration) {
       buffer.putUint8(137);
       writeValue(buffer, value.encode());
-    } else if (value is TexturePlayerIds) {
+    } else if (value is CreationOptions) {
       buffer.putUint8(138);
       writeValue(buffer, value.encode());
-    } else if (value is PlaybackState) {
+    } else if (value is TexturePlayerIds) {
       buffer.putUint8(139);
       writeValue(buffer, value.encode());
-    } else if (value is AudioTrackMessage) {
+    } else if (value is PlaybackState) {
       buffer.putUint8(140);
       writeValue(buffer, value.encode());
-    } else if (value is ExoPlayerAudioTrackData) {
+    } else if (value is AudioTrackMessage) {
       buffer.putUint8(141);
       writeValue(buffer, value.encode());
-    } else if (value is NativeAudioTrackData) {
+    } else if (value is ExoPlayerAudioTrackData) {
       buffer.putUint8(142);
       writeValue(buffer, value.encode());
-    } else if (value is ExoPlayerVideoTrackData) {
+    } else if (value is NativeAudioTrackData) {
       buffer.putUint8(143);
       writeValue(buffer, value.encode());
-    } else if (value is NativeVideoTrackData) {
+    } else if (value is ExoPlayerVideoTrackData) {
       buffer.putUint8(144);
+      writeValue(buffer, value.encode());
+    } else if (value is NativeVideoTrackData) {
+      buffer.putUint8(145);
       writeValue(buffer, value.encode());
     } else {
       super.writeValue(buffer, value);
@@ -910,20 +960,22 @@ class _PigeonCodec extends StandardMessageCodec {
       case 136:
         return PlatformVideoViewCreationParams.decode(readValue(buffer)!);
       case 137:
-        return CreationOptions.decode(readValue(buffer)!);
+        return PlatformWidevineDrmConfiguration.decode(readValue(buffer)!);
       case 138:
-        return TexturePlayerIds.decode(readValue(buffer)!);
+        return CreationOptions.decode(readValue(buffer)!);
       case 139:
-        return PlaybackState.decode(readValue(buffer)!);
+        return TexturePlayerIds.decode(readValue(buffer)!);
       case 140:
-        return AudioTrackMessage.decode(readValue(buffer)!);
+        return PlaybackState.decode(readValue(buffer)!);
       case 141:
-        return ExoPlayerAudioTrackData.decode(readValue(buffer)!);
+        return AudioTrackMessage.decode(readValue(buffer)!);
       case 142:
-        return NativeAudioTrackData.decode(readValue(buffer)!);
+        return ExoPlayerAudioTrackData.decode(readValue(buffer)!);
       case 143:
-        return ExoPlayerVideoTrackData.decode(readValue(buffer)!);
+        return NativeAudioTrackData.decode(readValue(buffer)!);
       case 144:
+        return ExoPlayerVideoTrackData.decode(readValue(buffer)!);
+      case 145:
         return NativeVideoTrackData.decode(readValue(buffer)!);
       default:
         return super.readValueOfType(type, buffer);